+++ src/SharpLink.Runtime/AssemblyInternals.cs
@@ -1,4 +1,5 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("SharpLink.Benchmarks")]
+[assembly: InternalsVisibleTo("SharpLink.Client")]
 [assembly: InternalsVisibleTo("SharpLink.UnitTests")]